Detailed weather screen done. Dev work completed

+++ app/src/main/java/com/tlc/weather/app/model/WindData.kt
@@ -1,3 +1,3 @@
 package com.tlc.weather.app.model
 
-data class WindData(val speed: Int, val deg: Int)
+data class WindData(val speed: Float, val deg: Float)
